pyproject.toml
[tool.poetry]
name = "campaign-zip-code-sync"
version = "0.1.0"
description = ""
authors = ["Charles Bannister <cbannister@godaddy.com>"]
readme = "README.md"

packages = [{ include = "zip_sync" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pytest = "^8.4.1"
load-dotenv = "^0.1.0"
oauth2client = "^4.1.3"
google-ads = "^27.0.0"
pandas = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
